Universal protobuf dump format (#604)
* [WIP] Universal protobuf dump format

* SMSG_CHAT to protobuf parsing

* SMSG_EMOTE to protobuf parsing

* SMSG_PLAY_SOUND to protobuf parsing

* SMSG_PLAY_MUSIC to protobuf parsing

* SMSG_PLAY_OBJECT_SOUND to protobuf parsing

* CMSG_GOSSIP_HELLO to protobuf parsing

* SMSG_GOSSIP_MESSAGE to protobuf parsing

* CMSG_GOSSIP_SELECT_OPTION to protobuf parsing

* CMSG_CLOSE_INTERACTION to protobuf parsing

* SMSG_SPELL_GO and SMSG_SPELL_START to protobuf parsing

* SMSG_AURA_UPDATE to protobuf parsing

* ON_MONSTER_MOVE to protobuf parsing

* SMSG_PHASE_SHIFT_CHANGE to protobuf parsing

* CMSG_SPELL_CLICK to protobuf parsing

* SMSG_QUERY_CREATURE_RESPONSE to protobuf parsing

* CMSG_PLAYER_LOGIN to protobuf parsing

* SMSG_PLAY_ONE_SHOT_ANIM_KIT to protobuf parsing

* SMSG_SET_AI_ANIM_KIT to protobuf parsing

* SMSG_PLAY_SPELL_VISUAL_KIT to protobuf parsing

* CMSG_QUEST_GIVER_ACCEPT_QUEST to protobuf parsing

* CMSG_QUEST_GIVER_COMPLETE_QUEST to protobuf parsing

* SMSG_QUEST_GIVER_QUEST_COMPLETE to protobuf parsing

* SMSG_QUEST_GIVER_REQUEST_ITEMS to protobuf parsing

* SMSG_QUERY_NPC_TEXT_RESPONSE to protobuf parsing

* SMSG_DB_REPLY to protobuf parsing

* SMSG_UPDATE_OBJECT to protobuf parsing

* Added protobuf structure version

* Better handling parsed proto file (locking/check if locked)

* SMSG_QUERY_GAME_OBJECT_RESPONSE to protobuf parsing

* CMSG_AREA_TRIGGER to protobuf parsing

* Added --ProtoVersion switch to get current proto version

* Added UniversalSplineFlags

* Parsing to protobuf will not write to txt file

* Extracted protobuf stuff to separate project

* Normalized protobuf fields names

* Added IPacketProcessor interface and base class

* Added movement data to protobuf UPDATE_OBJECT packet parsing

* Added additional data to SPELL_START and SPELL_GO proto structures

* Renamed 'W' component in proto Vector4 to 'O'

* Each update in UPDATE_OBJECT will now store its text in protobuf with text output mode

* Save bytes updates to proto

* SMSG_QUERY_PLAYER_NAME_RESPONSE to protobuf parsing

* Added missing new lines at the end

+++ WowPacketParser/Loading/SniffFile.cs
@@ -7,11 +7,14 @@ using System.IO.Compression;
 using System.Linq;
 using System.Text;
 using System.Threading;
+using Google.Protobuf;
 using WowPacketParser.Enums;
 using WowPacketParser.Enums.Version;
 using WowPacketParser.Hotfix;
 using WowPacketParser.Misc;
+using WowPacketParser.PacketStructures;
 using WowPacketParser.Parsing;
+using WoWPacketParser.Proto;
 using WowPacketParser.Saving;
 using WowPacketParser.SQL;
 using WowPacketParser.Store;
@@ -138,19 +141,38 @@ namespace WowPacketParser.Loading
                 case DumpFormatType.SqlOnly:
                 case DumpFormatType.Text:
                 case DumpFormatType.HexOnly:
+                case DumpFormatType.UniversalProto:
+                case DumpFormatType.UniversalProtoWithText:
                 {
                     var outFileName = Path.ChangeExtension(FileName, null) + "_parsed.txt";
+                    var outProtoFileName = Path.ChangeExtension(FileName, null) + "_parsed.dat";
+                    FileStream protoOutputStream = null;
 
-                    if (Utilities.FileIsInUse(outFileName) && Settings.DumpFormat != DumpFormatType.SqlOnly)
+                    if (Settings.DumpFormatWithTextToFile())
                     {
-                        // If our dump format requires a .txt to be created,
-                        // check if we can write to that .txt before starting parsing
-                        Trace.WriteLine($"Save file {outFileName} is in use, parsing will not be done.");
-                        break;
+                        if (Utilities.FileIsInUse(outFileName) && Settings.DumpFormat != DumpFormatType.SqlOnly)
+                        {
+                            // If our dump format requires a .txt to be created,
+                            // check if we can write to that .txt before starting parsing
+                            Trace.WriteLine($"Save file {outFileName} is in use, parsing will not be done.");
+                            break;
+                        }
+                        File.Delete(outFileName);
                     }
-
+                    
+                    if (_dumpFormat is DumpFormatType.UniversalProto or 
+                        DumpFormatType.UniversalProtoWithText)
+                    {
+                        if (Utilities.FileIsInUse(outProtoFileName))
+                        {
+                            Trace.WriteLine($"Save file {outProtoFileName} is in use, parsing will not be done.");
+                            break;
+                        }
+                        File.Delete(outProtoFileName);
+                        protoOutputStream = File.Create(outProtoFileName);
+                    }
+                    
                     Store.Store.SQLEnabledFlags = Settings.SQLOutputFlag;
-                    File.Delete(outFileName);
 
                     _stats.SetStartTime(DateTime.Now);
 
@@ -161,8 +183,10 @@ namespace WowPacketParser.Loading
                     ThreadPool.SetMinThreads(threadCount + 2, 4);
 
                     var written = false;
-                    using (var writer = (Settings.DumpFormatWithText() ? new StreamWriter(outFileName, true) : null))
+
+                    using (var writer = (Settings.DumpFormatWithTextToFile() ? new StreamWriter(outFileName, true) : null))
                     {
+                        Packets packets = new() { Version = StructureVersion.ProtobufStructureVersion };
                         var firstRead = true;
                         var firstWrite = true;
 
@@ -202,6 +226,8 @@ namespace WowPacketParser.Loading
                         {
                             if (!Console.IsOutputRedirected)
                                 ShowPercentProgress("Processing...", reader.PacketReader.GetCurrentSize(), reader.PacketReader.GetTotalSize());
+                            else
+                                Console.WriteLine(reader.PacketReader.GetCurrentSize() * 1.0 / reader.PacketReader.GetTotalSize());
 
                             if (!packet.Status.HasAnyFlag(Settings.OutputFlag) || !packet.WriteToFile)
                             {
@@ -246,18 +272,31 @@ namespace WowPacketParser.Loading
                             }
 // ReSharper restore AccessToDisposedClosure
 
+                            if (_dumpFormat is DumpFormatType.UniversalProtoWithText)
+                                packet.Holder.BaseData.StringData = packet.Writer.ToString();
+
                             // Close Writer, Stream - Dispose
                             packet.ClosePacket();
+                            
+                            if (_dumpFormat is DumpFormatType.UniversalProto or 
+                                DumpFormatType.UniversalProtoWithText)
+                                packets.Packets_.Add(packet.Holder);
                         }, threadCount);
 
                         pwp.WaitForFinished(Timeout.Infinite);
 
                         reader.PacketReader.Dispose();
 
+                        if (protoOutputStream != null)
+                        {
+                            packets.WriteTo(protoOutputStream);
+                            protoOutputStream.Close();
+                        }
+
                         _stats.SetEndTime(DateTime.Now);
                     }
 
-                    if (Settings.DumpFormatWithText())
+                    if (Settings.DumpFormatWithTextToFile())
                     {
                         if (written)
                             Trace.WriteLine($"{_logPrefix}: Saved file to '{outFileName}'");
